chore(dependency): remove eigen library

+++ projects/frontend/es-app/src/CommandThread.cpp
@@ -83,7 +83,7 @@ void CommandThread::run() {
 void CommandThread::runGame (FileData* game)
 {
 	ViewController *view = ViewController::get();
-	Eigen::Vector3f target(Renderer::getScreenWidth() / 2.0f, Renderer::getScreenHeight() / 2.0f, 0);
+	Vector3f target((float)Renderer::getScreenWidth() / 2.0f, (float)Renderer::getScreenHeight() / 2.0f, 0);
 	mWindow->doWake();
 	view->launch(game, target);
 }
